Changed the architecture of LockEngine. Now, the responsibilities of LockEngine and ClassMolder are clearly separated.
LockEngine is responsible for maintaining locks and concurrence control.

LockEngine invokes ClassMolder and provides thread safe environment
for ClassMolder. An object, DepositBox, will be passed by LockEngine to
ClassMolder. ClassMolder can put the dirty checking cache in DepositBox.
Only the transaction in action has the proper accessMode will be let to
access the DepositBox. If the transaction in action has the writeLock,
it can set and read the object. If the transaction in action has only
readLock, it can only read the object.

With one exception: the transaction is trying to load an object with
readMode but the object doesn't exist in cache (getObject return null).
In such situation, the transaction may call setObject once. And, it is
guaranteed no other thread interested into the same lock will let running
concurrently with the loading transaction. So, it can safely call
PersistenceEngine and put the result into DepositBox.

ClassMolder can also obtain TimeStamp from the DepositBox. Which will
be reset to the system time, whenever setObject is invoked on the
DepositBox.

ClassMolder is responsible for the actual object creation, deletion,
loading, updating and persisting. Beside interacting with DepositBox,
ClassMolder interacts with TransactionContext, to resolve relation
to other object and update its accessMode if necessary.