Added checking mechanism to ensure LRU cache consistency for related objects. when a object is stored in perpare state. If an object is deleted, clean up will be done to the related object. So, the relationship will deleted in the referring object as well. Also, when an object is stored, if a relation is deleted, clean up will be done to the related object to make sure that object is not holding reference to itself.
Also, store is splited into two methods. store() and
preStore(). preStore will be responsible the above newly
added checking as well as all other existing dependent
object creation, removal, and Many-to-many relation
bookkeeping etc. It may trigger other markDelete etc.

And, store will be called after all markDeleted and
preStore is done to all objects in the transaction. And,
it will persist the change to the persistence storage.