ClassMolder's Store is splited into two methods. store() and preStore(). preStore will be responsible the above newly added checking as well as all other existing dependent object creation, removal, and Many-to-many relation bookkeeping etc. It may trigger other markDelete etc.
And, store will be called after all markDeleted and
preStore is done to all objects in the transaction. And,
it will persist the change to the persistence storage.
(See also ClassMolder)

The split ensures that all changes to a object won't be
pesisted more than one in prepare state which may result
from the newly added bookkeeping on related objects if
the split isn't done.

Added markModified mechanism. And, transaction is now
distingish between cache updated need and persistence
storage updated needed.

Fixed a bug which will recreated an new instance of deleted
object.