Fixed a bug with many-many relation which insert (null, null) into relation SQL table, when key generators are used on both side.
Delayed the creation of dependent object to the preStore state. It
avoid problems if key generators are used.

Update cache at the end of create method to make sure the cache
contains the latest newly generated keys.

Updated the method of getAddedValueList. Object with null identity
is considered as newly added.