Fixed a potential problem with lazy collection consistency.
If an object is removed from database, and the lazy collection
still link to it at the end of transaction, the cache would
still "think" the object is in the database in the next
transaction.

Although the test can detect that in some case, it is not
recommended to depends on this behavior. Always dis-link it,
before calling db.remove().